ML/requirements.txt
uvicorn 
fastapi
scholarly
kaggle
groq
python-dotenv
pydantic
